Add WriteKey and deprecate TeamId (#30)
Fixes #14.

Adds HoneycombApiSettings.WriteKey to replace TeamId to improve consistency between beeline implementations. TeamId is marked as deprecated and any usage manipulates the new property to ensure backwards compatibility. Docs and usage examples also updated.

Also includes replacing package references in Honeycomb.AspnetCore and the Sample projects with project references so it uses the local projects instead of depending on downloaded packages. This allows easier code debugging and for any changes in the root project to be reflected in sub-projects without having to release the main project first.