Prepare v1.1.1. release (#32)
Updates package versions to v1.1.1 and adds a Changelog (includes entries for v1.1.0 & v1.1.1).

Also removes the VersionSuffix property from csproj as they were causing a problem with CI building the packages. We don't really need it anyway, it was a nice to have to separate local builds from official packages.